+++ src/mgfw/math/Vector4.h
@@ -49,8 +49,8 @@ struct Vector4 {
 
     /** Operators */
 
-    T& operator[]( uint16_t index ) { return data[ index ]; }
-    const T& operator[]( uint16_t index ) const { return data[ index ]; }
+    T& operator[]( uint8_t index ) { return data[ index ]; }
+    const T& operator[]( uint8_t index ) const { return data[ index ]; }
 
     template <typename U>
     bool operator==( const Vector4<U>& other ) const {
